requirements.txt
Flask==2.3.3
girth==0.8.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
PasteScript==3.6.0
Pillow==10.3.0
scikit_learn==1.4.2
waitress==2.1.2
PyLTI1p3==2.0.0
pycryptodome==3.20.0
Flask-Caching==2.3.0
Flask-Session2==1.3.1
Flask-CORS==5.0.0
Werkzeug>=3.0.3
cachelib==0.9.0
pinecone-client==5.0.1
Pinecone == 5.3.1
openai==1.55.3
canvasapi==2.2.0
python-dotenv==1.0.1
python-memcached==1.62

